fix: address PR #9683 review feedback
- Inline extractFunctionToolCallFields back into conversationFormat.ts as
  a local, non-exported function. It had exactly one caller
  (formatTopLevelToolCall); normalizeConversation.ts's getAssistantToolCalls
  keeps its own separate SDK-narrowed extraction and never called it, so
  the shared module's claim of sharing it was wrong and the extraction
  itself violated the single-caller-extraction rule.

- Read followupOptions through a new memoized streamContext$ selector
  (mirroring streamById$/streamStates$'s memoizedFacetView pattern)
  instead of the raw streamEntries$ signal. streamEntries$ changes
  identity on any per-stream write, including log-only ticks on any
  stream, so streamContext$ was recomputing (and its non-log consumers
  re-rendering) on every streamed token — a regression versus the
  pre-consolidation followupOptionsByStream map, which lived outside
  streamLogs and wasn't entangled with log writes.

- Reset the two facet-memoization caches (_prevStreamById/_prevStreamStates)
  and the new followupOptions one in resetProgressState(), matching the
  existing _prevApprovalIds/_prevPhaseStages handling. Self-healing on the
  next read either way, but kept for consistency with the pattern this
  file documents.
